+++ project/lib/models/dataBigWidget.dart
@@ -1,7 +1,7 @@
 class dataBigWidget{
-  int data1;
-  int data2;
-  int data3;
+  double data1;
+  double data2;
+  double data3;
 
   dataBigWidget({required this.data1,required this.data2,required this.data3});
 }
