+++ project/lib/models/transections.dart
@@ -1,11 +1,17 @@
 // ignore: camel_case_types
-class transections{
-  int dataMeat;
-  int dataRice;
-  int dataVeget;
-  int dataFruit;
-  int dataMilk;
+class transections {
+  double dataMeat;
+  double dataRice;
+  double dataVeget;
+  double dataFruit;
+  double dataMilk;
   DateTime date;
 
-  transections({required this.dataMeat,required this.dataRice,required this.dataVeget,required this.dataFruit,required this.dataMilk,required this.date});
+  transections(
+      {required this.dataMeat,
+      required this.dataRice,
+      required this.dataVeget,
+      required this.dataFruit,
+      required this.dataMilk,
+      required this.date});
 }
